Use new v0.16 API for creating a multisig address
This doesn't work with v0.15.1 or any older version of Bitcoin
Core. I'll add a version check soon so we don't get an ugly error
message if anyone tries it.

Doesn't actually change anything since I'm using 'legacy' as the
address type, but this will enable a switch to p2wsh-in-p2sh (segwit)
soon.