Revert "Teach wallet about cold storage address during withdrawal"
During review, we realized this will write the privkeys into the
wallet.dat, which opens up a new potential attack vector. To be safe,
let's not do that.

The bug this was intended to work around was fixed before the release
of Bitcoin Core v0.16 so it should no longer be necessary. I will find
another way to fix #20.

This reverts commit 1d2951faf194bfe3ae1f077a824aae0599904939.

Conflicts:
	glacierscript.py