Recreate developer tests' golden data using Bitcoin Core v0.17.0rc1
Also reran t/decode-transactions to demonstrate and document the
changes to the constructed withdrawal transactions.

Because of https://github.com/bitcoin/bitcoin/pull/13666 the signature
sizes have changed, which causes the transaction sizes to change,
which causes the fee calculations to change, which causes the values
in the transaction outputs to change. In every case where it changed,
the fee is less and the output is more.

Previously, the tests all passed using Bitcoin Core v0.16.2 and failed
on v0.17.0rc1. Now, the tests pass on v0.17.0rc1 and fail on v0.16.2.

This commit also shows evidence of issue #36 ("WARNING: transaction QR
code could not be verified properly") in
create-withdrawal-data.uncompressed.golden. I will be fixing that
false alarm soon.