Accept variable number of args to run_subprocess
This is preparation for having the callers provide parameters as a
list, instead of as one long string.

The **kwargs is unfortunate; if we had Python 3 I would have used
something like:

def run_subprocess(sub_func, exe, *args, *, silent=False):

The **kwargs technique for Python 2 is from the book "Effective
Python".