Add output for -v verbose
There's a couple of issues with this. (1) The "output" is not always
the output; sometimes it's the return code. (2) Arguments are not
properly quoted for shell escaping; one cannot simply cut & paste this
line into the shell to rerun. (3) When silent=True the output of the
command is lost. We'd like for -v to show that output for debug
purposes.