Sort list of developer tests before assigning RPC ports
This will help make the tests more stable from one run to the
next. Tests are more likely to run with the same port number they ran
with last time. That will help with the occasional test failures I get
from a stale `bitcoind` process having to be explicitly killed or else
the test will fail since bitcoind can't get a lock on the wallet.

The Makefile does try to kill `bitcoind` before and after every test,
but if the previous run failed, its bitcoind will still be running. If
on a subsequent run, its RPC port changes, the Makefile will not be
able to kill it, leading to test failures due to timeout.