Update decode-transactions to deal with new case
create-withdrawal-data.fails.run deliberately fails to create a
transaction, which was confusing decode-transactions since it expected
every create-withdrawal-data.*.golden to create a transaction at the
end.

Now, decode-transactions will point out that no transaction was found,
so that it's clear what happened, and if through some future bug the
failing case did actually create a transaction, we can spot the
difference.